Removed templateness of ReadWriteBufferFromS3.
